refactor: auth 패키지 java 컨벤션 적용
