+++ src/main/java/nextstep/auth/JwtTokenProvider.java
@@ -8,31 +8,32 @@ import java.util.Date;
 
 @Component
 public class JwtTokenProvider {
+
     private final String secretKey = "learning-test-spring";
     private final Long validityInMilliseconds = 3600000L;
 
     public String createToken(String principal) {
         Claims claims = Jwts.claims()
-                .setSubject(principal);
+            .setSubject(principal);
         Date now = new Date();
         Date validity = new Date(now.getTime() + validityInMilliseconds);
 
         return Jwts.builder()
-                .setClaims(claims)
-                .setIssuedAt(now)
-                .setExpiration(validity)
-                .signWith(SignatureAlgorithm.HS256, secretKey)
-                .compact();
+            .setClaims(claims)
+            .setIssuedAt(now)
+            .setExpiration(validity)
+            .signWith(SignatureAlgorithm.HS256, secretKey)
+            .compact();
     }
 
     public String getPrincipal(String token) {
         String parsedToken = null;
         try {
             parsedToken = Jwts.parser()
-                    .setSigningKey(secretKey)
-                    .parseClaimsJws(token)
-                    .getBody()
-                    .getSubject();
+                .setSigningKey(secretKey)
+                .parseClaimsJws(token)
+                .getBody()
+                .getSubject();
         } catch (JwtException jwtException) {
             throw new JwtException("유효하지 않은 토큰입니다.");
         } catch (IllegalArgumentException illegalArgumentException) {
