+++ src/main/java/nextstep/auth/utils/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package nextstep.infrastructure;
+package nextstep.auth.utils;
 
 import nextstep.support.exception.NoAccessTokenException;
 
@@ -6,8 +6,10 @@ import javax.servlet.http.HttpServletRequest;
 import java.util.Enumeration;
 
 public class AuthorizationExtractor {
+
     public static final String AUTHORIZATION = "Authorization";
-    public static final String ACCESS_TOKEN_TYPE = AuthorizationExtractor.class.getSimpleName() + ".ACCESS_TOKEN_TYPE";
+    public static final String ACCESS_TOKEN_TYPE =
+        AuthorizationExtractor.class.getSimpleName() + ".ACCESS_TOKEN_TYPE";
     public static String BEARER_TYPE = "Bearer";
 
     public static String extract(HttpServletRequest request) {
@@ -17,12 +19,9 @@ public class AuthorizationExtractor {
         }
         while (headers.hasMoreElements()) {
             String value = headers.nextElement();
-            if ((value.toLowerCase()
-                    .startsWith(BEARER_TYPE.toLowerCase()))) {
-                String authHeaderValue = value.substring(BEARER_TYPE.length())
-                        .trim();
-                request.setAttribute(ACCESS_TOKEN_TYPE, value.substring(0, BEARER_TYPE.length())
-                        .trim());
+            if ((value.toLowerCase().startsWith(BEARER_TYPE.toLowerCase()))) {
+                String authHeaderValue = value.substring(BEARER_TYPE.length()).trim();
+                request.setAttribute(ACCESS_TOKEN_TYPE, value.substring(0, BEARER_TYPE.length()).trim());
                 int commaIndex = authHeaderValue.indexOf(',');
                 if (commaIndex > 0) {
                     authHeaderValue = authHeaderValue.substring(0, commaIndex);
